refactor: Apply RSAPE
